Google Earth provider changes
Google Earth relies on a callback before any changes are made to the
map. Custom attributes can not be added to Google Earth objects like
mapstraction_marker - perhaps the id of the element added can be used to
link mapstraction and earth elements?
This provider had dependencies on GEarthExtensions and geo, I could find
no way of setting the map bounds without this